formatter: fix issue where formatter may not register properly
Previously, the formatter was only registered if the callback for `exec`
happened to execute prior to the check for the install, which I couldn't
get this to happen consistently on my machine.

To fix this, registering the formatter is now done in the exec callback.

I've made some minor other style tweaks as well.